+++ doc/build.info
@@ -675,6 +675,10 @@ DEPEND[html/man3/BIO_set_callback.html]=man3/BIO_set_callback.pod
 GENERATE[html/man3/BIO_set_callback.html]=man3/BIO_set_callback.pod
 DEPEND[man/man3/BIO_set_callback.3]=man3/BIO_set_callback.pod
 GENERATE[man/man3/BIO_set_callback.3]=man3/BIO_set_callback.pod
+DEPEND[html/man3/BIO_set_dgram_origin.html]=man3/BIO_set_dgram_origin.pod
+GENERATE[html/man3/BIO_set_dgram_origin.html]=man3/BIO_set_dgram_origin.pod
+DEPEND[man/man3/BIO_set_dgram_origin.3]=man3/BIO_set_dgram_origin.pod
+GENERATE[man/man3/BIO_set_dgram_origin.3]=man3/BIO_set_dgram_origin.pod
 DEPEND[html/man3/BIO_should_retry.html]=man3/BIO_should_retry.pod
 GENERATE[html/man3/BIO_should_retry.html]=man3/BIO_should_retry.pod
 DEPEND[man/man3/BIO_should_retry.3]=man3/BIO_should_retry.pod
@@ -2896,6 +2900,7 @@ html/man3/BIO_s_mem.html \
 html/man3/BIO_s_null.html \
 html/man3/BIO_s_socket.html \
 html/man3/BIO_set_callback.html \
+html/man3/BIO_set_dgram_origin.html \
 html/man3/BIO_should_retry.html \
 html/man3/BIO_socket_wait.html \
 html/man3/BN_BLINDING_new.html \
@@ -3490,6 +3495,7 @@ man/man3/BIO_s_mem.3 \
 man/man3/BIO_s_null.3 \
 man/man3/BIO_s_socket.3 \
 man/man3/BIO_set_callback.3 \
+man/man3/BIO_set_dgram_origin.3 \
 man/man3/BIO_should_retry.3 \
 man/man3/BIO_socket_wait.3 \
 man/man3/BN_BLINDING_new.3 \
