+++ util/other.syms
@@ -160,6 +160,8 @@ BIO_get_buffer_num_lines                define
 BIO_get_cipher_ctx                      define
 BIO_get_cipher_status                   define
 BIO_get_close                           define
+BIO_get_dgram_dest                      define
+BIO_get_dgram_origin                    define
 BIO_get_ktls_send                       define
 BIO_get_ktls_recv                       define
 BIO_get_conn_address                    define
@@ -201,6 +203,8 @@ BIO_set_conn_hostname                   define
 BIO_set_conn_port                       define
 BIO_set_conn_ip_family                  define
 BIO_set_conn_mode                       define
+BIO_set_dgram_dest                      define
+BIO_set_dgram_origin                    define
 BIO_set_fd                              define
 BIO_set_fp                              define
 BIO_set_indent                          define
